pyproject.toml
[tool.poetry]
name = "photo-organization-tool"
version = "0.1.0"
description = ""
authors = ["m-shao <dev.minglunshao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
piexif = "^1.1.3"
pillow = "^10.4.0"
pyexiftool = "^0.5.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
